Use events instead of mixin for paw restrictions
Also split PAWS_ALLOW_INTERACT into two (one for break, one for
interact)
