+++ src/main/java/org/jlortiz/playercollars/PlayerCollarsMod.java
@@ -60,6 +60,7 @@ import org.jetbrains.annotations.Nullable;
 import org.jlortiz.playercollars.block.DogBedBlock;
 import org.jlortiz.playercollars.block.DogBowlBlock;
 import org.jlortiz.playercollars.block.InvisibleFenceBlock;
+import org.jlortiz.playercollars.event.PawsEventHandler;
 import org.jlortiz.playercollars.item.*;
 import org.jlortiz.playercollars.leash.LeashImpl;
 import org.jlortiz.playercollars.leash.LeashProxyEntity;
@@ -158,6 +159,7 @@ public class PlayerCollarsMod implements ModInitializer {
 	public static final DyeColor[] PAWS_DYE_COLORS = new DyeColor[]{DyeColor.WHITE, DyeColor.LIGHT_GRAY,
 			DyeColor.GRAY, DyeColor.BLACK, DyeColor.BLUE, DyeColor.RED, DyeColor.PURPLE};
 	public static final PawsItem[] PAWS_ITEMS = new PawsItem[PAWS_DYE_COLORS.length];
+	public static final TagKey<Block> PAWS_ALLOW_BREAK = TagKey.of(RegistryKeys.BLOCK, Identifier.of(MOD_ID, "paws_allow_break"));
 	public static final TagKey<Block> PAWS_ALLOW_INTERACT = TagKey.of(RegistryKeys.BLOCK, Identifier.of(MOD_ID, "paws_allow_interact"));
 	public static final TagKey<Item> PAWS_TAG = TagKey.of(RegistryKeys.ITEM, Identifier.of(MOD_ID, "paws"));
 	public static final FootPawsItem[] FOOT_PAWS_ITEMS = new FootPawsItem[PAWS_DYE_COLORS.length];
@@ -389,5 +391,7 @@ public class PlayerCollarsMod implements ModInitializer {
 			if (entity instanceof LeashKnotEntity ke && blockLeashKnotBreak(sworld, player, ke)) return ActionResult.FAIL;
 			return ActionResult.PASS;
 		});
+
+		PawsEventHandler.registerPawsEvents();
 	}
 }
