+++ src/main/resources/playercollars.mixins.json
@@ -12,8 +12,5 @@
     "LivingEntityMixin",
     "PlayerEntityMixin",
     "ServerPlayerEntityMixin"
-  ],
-  "client": [
-    "ClientPlayerInteractionManagerMixin"
   ]
 }
